Move scripts around, fix travis and dockerfile
