+++ plankton/__main__.py
@@ -17,98 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # *********************************************************************
 
-import argparse
-import os
-import sys
-
-from plankton.adapters import import_adapter, get_available_adapters
-from plankton.core.control_server import ControlServer, ExposedObject
-from plankton.core.utils import get_available_submodules
-from plankton.core.version import __version__
-from plankton.devices import import_device
-
-from plankton.core.simulation import Simulation
-
-
-def main():
-    parser = argparse.ArgumentParser(
-        description='Run a simulated device and expose it via a specified communication protocol.')
-
-    parser.add_argument('-r', '--rpc-host',
-                        help='HOST:PORT format string for exposing the device via JSON-RPC over ZMQ.')
-    parser.add_argument('-s', '--setup', default=None,
-                        help='Name of the setup to load.')
-    parser.add_argument('-l', '--list-protocols',
-                        help='List available protocols for selected device.', action='store_true')
-    parser.add_argument('-p', '--protocol', default=None,
-                        help='Communication protocol to expose devices.')
-    parser.add_argument('-c', '--cycle-delay', type=float, default=0.1,
-                        help='Approximate time to spend in each cycle of the simulation. '
-                             '0 for maximum simulation rate.')
-    parser.add_argument('-e', '--speed', type=float, default=1.0,
-                        help='Simulation speed. The actually elapsed time between two cycles is '
-                             'multiplied with this speed to determine the simulated time.')
-    parser.add_argument('-k', '--device-package', default='plankton.devices',
-                        help='Name of packages where devices are found.')
-    parser.add_argument('-a', '--add-path', default=None,
-                        help='Path where the device package exists. Is added to the path.')
-    parser.add_argument('-v', '--version', action='store_true',
-                        help='Prints the version and exits.')
-
-    parser.add_argument('device', nargs='?',
-                        help='Name of the device to simulate, '
-                             'omitting prints list of available devices.')
-    parser.add_argument('adapter_args', nargs='*',
-                        help='Arguments for the adapter.')
-
-    arguments = parser.parse_args()
-
-    if arguments.version:
-        print(__version__)
-        return
-
-    if arguments.add_path is not None:
-        sys.path.append(os.path.abspath(arguments.add_path))
-
-    if not arguments.device:
-        print('Please specify a device to simulate.')
-        print('The following devices are available:')
-
-        for dev in get_available_submodules(arguments.device_package):
-            print('\t' + dev)
-        return
-
-    # Import the device type and required initialisation parameters.
-    device_type, parameters = import_device(arguments.device, arguments.setup,
-                                            device_package=arguments.device_package)
-
-    if arguments.list_protocols:
-        adapters = get_available_adapters(arguments.device, device_package=arguments.device_package)
-
-        protocols = {adapter.protocol for adapter in adapters.values()}
-
-        for p in protocols:
-            print(p)
-        return
-
-    device = device_type(**parameters)
-    adapter = import_adapter(arguments.device, arguments.protocol,
-                             device_package=arguments.device_package)(device, arguments.adapter_args)
-
-    simulation = Simulation(
-        device=device,
-        adapter=adapter)
-
-    simulation.cycle_delay = arguments.cycle_delay
-    simulation.speed = arguments.speed
-
-    if arguments.rpc_host:
-        simulation.control_server = ControlServer(
-            {'device': device,
-             'simulation': ExposedObject(simulation, exclude=('start', 'control_server'))},
-            *arguments.rpc_host.split(':'))
-
-    simulation.start()
+from plankton.scripts.run import run_simulation
 
 if __name__ == '__main__':
-    main()
+    run_simulation()
